feat: counttable migration
